Addl fix #397: GPU acceleration
https://github.com/BestImageViewer/geeqie/issues/397

If two windows are open, switching between the two windows may cause a
crash with the message:
The error was 'GLXBadDrawable'

The bug happens with Nouveau drivers and Radeon drivers, but not with
Nvidia drivers.

This patch disables Clutter if the system is liable to this bug. The
user is given the possibility to override this with an option in
Preferences/Behavior.