Fix #822: The image rotation keys ([ and ]) affect the wrong image
https://github.com/BestImageViewer/geeqie/issues/822

If a call is of the format:
geeqie /<folder1>/*jpg
Geeqie opens folder1 with the listed files selected.

If a call is of the format:
geeqie /<folder1>/*jpg /<folder2>/*jpg
Geeqie opens folder1 and the listed files are displayed in an unnamed
collection.

The hidden command line collection feature of previous versions is
deleted.