Roughly splits filedata.c into multiple files that are grouped around general functionalities.
This relies on treating the files as C++, which allows them to share private APIs that are inaccessible to other callers, even though they're now in separate compilation units.

This is a first step towards establishing a well-defined API for FileData, and organizing the code in a way that is easier to understand and easier to validate.  The eventual goal would be to simplify the FileData core, and use inheritance or composition to layer the more complex functionalities in a way that doesn't make the base implementation as extraordinarily complex as it's become.