fix unpredictable behavior in doxygen generator script
We know that SRCDIR is the location of doxygen.conf, and that it is ../
from the doc directory. But we don't know that the current working
directory of the script is the doc directory -- the reference version of
Meson currently sets that, but other implementations of Meson such as
muon (c99) set it to the project root. Meson's own documentation says to
NOT rely on this as it is officially unspecified.

The solution is actually super simple though, because we already pass
the project root as a script argument, so just... use that. :D