AppImage version update notification
If Geeqie is being run as an AppImage and the version on the server is
newer than the one currently being run, a fade-out message is shown on
start-up.
This can be disabled via an option on Preferences/General.
An Internet connection is required.
The option is not displayed if not being run as an AppImage.