Modify part of Collections user interface
Collections .gqv files are expected to be in the default directory.
The user no longer sees file paths or file extensions, only the
collection name.
If the user navigates to a .gqv file located elsewhere and opens it, the
contents will be displayed and operated on as a collection. It will
appear in the "recently used collections" menu (there may be a name
conflict) but not on the "open collection" menu, which shows only the
collections in the default directory.