Fix #1185: Do you plan to embed video player like gthumb or pix has their own built-in player?
https://github.com/BestImageViewer/geeqie/issues/1185

Implement a plugin that uses xdg-open to run an external video player on
video files.

New users will have class video files selected, the "Play video with
left click" selected, and the "Play with" option set to the the new
plugin.

Existing users will need to select the options themselves.