Bug fix: Collections and unmounted drives
The bug:
If a collection file is on the local filesystem
The collection file contains references to files on a mounted drive
The drive is unmounted
The collection file is opened

Geeqie sees that the files do not exist and deletes them from the
collection file.

The fix:
If Geeqie does not find the file-

If the file path prefix contains /home, /tmp or /usr it is assumed that
the file was on the local drive and has been deleted by the user. It is
ignored.

If the file path prefix contains a mount path, it is assumed that the
file was on a currently mounted drive and has been deleted.

Otherwise it is assumed that the file is on a removable drive that is
not currently mounted. The collection will not be opened.

If this is not the case the user may need to use a text editor to remove
the offending line from the collection file.

Created problem:
When a collection window contents have not been changed, but its
geometry has, the geometry changes will not be saved. The user must make
a minor positional change to any thumbnail to trigger a save.