DO NOT MERGE Creates an early prototype of one way to implement unit tests in Geeqie
Last year, I had tried to untangle the code enough to use googletest in a more
standard way (where it only compiles and links against the specific file under
test and avoids including the rest of Geeqie), and that didn't work because the
Geeqie codebase is so interdependent in its current state.

This approach comes from the opposite direction: just make a version of Geeqie
that runs the unit tests instead of actually starting the application.  The next
step would be to tie this into some sort of unit test mode flag in meson, and to
figure out how to incorporate some C++ unit test framework (I'm most familiar
with googletest, so that's my preference).

The meson config should also only ever compile and link the unit test code in
the unit test mode, so that code should not be shipped to users as part of
releases.

Lastly, it's typical to create parallel `src/` and `tests/` directory
hierarchies.  In this case, I think it'll be more straightforward to create
a `src/tests/` subdirectory and to put everything there.

```
$./build/src/geeqie
Unit test run function
A condition passed: util_clip_region found overlap
A condition passed: rx as expected
A condition passed: ry as expected
A condition passed: rw as expected
A condition passed: rh as expected
A condition failed: expected to fail
```