Bug fix: Possible seg. fault on exit
view-dir.cc folder_icons_new() sets up a set of pixbufs.
There is the possibility one or more may be initialized null.

On exit, folder_icons_free() is called, and will seg.fault if any of the
items are null.

This does not happen when running an a local computer, but it does occur
when running tests via a GitHub Action.

Check if the object is null before un-reffing.

[The tests were run with G_DEBUG=fatal-warnings]