Include the Help file in pdf format (2)
At the moment the help.pdf file is generated outside the Geeqie project
and uploaded to geeqie.org along with the other help files.

This is done on https://github.com/geeqie/geeqie.github.io

The Help menu code looks for the local file in gq_helpdir, does not find
it and defaults to the copy on geeqie.org.

This will be corrected when the help.pdf file can be generated from the
Docbook xml files.