Fix #1315: trying to remove folder with completely hidden .thumbnails folder silently fails
https://github.com/BestImageViewer/geeqie/issues/1315

When attempting to delete a directory, if Preferences/File Filter/ is
enabled and any hidden files exist, the user will not see them in the
dialog and the program will refuse to delete the directory.

Temporarily disable the file filter when deleting a directory. The
hidden files will be displayed in the dialog and the directory can be
deleted.