Improve user information when config. file error detected
Previously if a config. file error was detected, the error messages were
sent to the command line and to the log window. Therefore it was
possible that the user was unaware of the error.

When a config. file error is detected a notification is issued and also
the message is sent to the command line/log window.

The user will see only one notification if there are multiple errors.
The user must explicitly dismiss the notification.
The notification has a button to show the log window, where all errors
are displayed.