Temporary fix: AppImages - Wayland
The linuxdeploy gtk plugin sets the default to X11, with the comment
that using Wayland causes a crash - although this comment was made some
years ago

Limited local testing under Wayland does not show a problem.

The relevant line is patched out before the AppImage is created. The
backend therefore defaults to whatever is being used.