Status bar expander symbols
The Sort and Zoom boxes on the Status Bar have expander symbols to
indicate to the user that they have drop-down menus.

An unknown change in gtk resulted in the symbols not being displayed.

This commit includes a call to gtk_button_set_always_show_image() which
again ensures that the symbols are displayed.