Collection files on unmounted drives
The original problem is described in commit cf67a444c22

When a file is moved or renamed, the file_data_notification function
runs and checks Collections to see if the file is included there.

If a Collection includes a file on an unmounted drive, a warning message
was issued. That required the user to acknowledge the message.

Change the warning message to a notification, which does not require
user input.