AppImage creation - symlinks
If a Geeqie Appimage is created in the standard way, when the executable
is called via a symlink, apprun-hooks/linuxdeploy-plugin-gtk.sh is not
found and geeqie --help output is incorrect.
See:
https://github.com/linuxdeploy/linuxdeploy/issues/303
https://github.com/linuxdeploy/linuxdeploy/issues/304

This was partially solved for Geeqie by editing AppRun when
./scripts/geeqie-download-appimage.sh was run.
Clearly that did not apply if the user did not use that script.

This patch solves the problem by using the option linuxdeploy
--custom-apprun=[AppRun path] to ensure the patch is applied in all
cases.