Fix various AppImage problems
- Reverse commit 398d752eaf - that did not work (symlinks).
- Update README to show user how to avoid symlink problem.
- x86 AppImages show runtime errors due to library linking problems
(gio, glib) when run on Ubuntu 25.04 - due to the GitHub runner using
24.04.
Solve the problem by running AppImage creation in a docker which runs
25.04.
Note that arm64 AppImages cannot be created in this way due to
docker/sudo/pguyot/arm-runner-action problems.
Presumably the AppImages created on 25.04 will not run on 24.04 systems.
- Change the x86 AppImage .ymls to use ubuntu-latest and ubuntu-rolling.