Fix shortcuts dialogs
The dialogs had a Name button at the same level as the Open/Cancel
buttons. This area is reserved for actions that terminate the window.

The Name button modifies the contents of the window, but does not
terminate it.

Move the Name button to the extra widgets area and make it a text entry.