Facilitate AppImages for each formal Release
Script to download AppImages from Continuous Build, rename the files to
remove the word "latest" and replace with the version number.

The files may then be uploaded to the Geeqie Latest Release area on
GitHub.

The script should be run after a new release has been uploaded to GitHub
and the new Continuous Build AppImages have been generated.