Move folder/file dialogs to gtk_file_chooser()
All folder/file dialogs are via gtk_file_chooser(). This provides a
consistent user interface, and the dialog itself has advantages.

Previews of folders/some images/pdf/icc/some text files.

A history combo is included where sensible.

Regression:

The copy/move dialog had a "dot" as a shortcut to the current folder.
The new dialog also has a shortcut to the current folder, but its name
is the final folder name in the path - there is no way to change it to
be a "dot". It is therefore not easy to see which the current folder
shortcut is.